Sofas and waiting room implemented
Worked from the ground up implementation. Follows example output scheme. Uses semaphores for sofa waiting, and mutex for entering waiting room. 
Note that neither of these counters are decremented as they will be decremented by other functions in other parts, so will need to add those later. 
MedProf funtion is not implemented but added, will also need to add correct struct that we agreed on using for these objects
More needs to be added and will be later.